+++ internal/config/common.go
@@ -1,6 +1,7 @@
 package config
 
 import (
+	"log/slog"
 	"time"
 
 	commonConfig "github.com/crypto-bundle/bc-wallet-common-lib-config/pkg/config"
@@ -45,4 +46,33 @@ type loggerCfgService interface {
 	GetMinimalLogLevel() string
 	IsStacktraceEnabled() bool
 	GetSkipBuildInfo() bool
+}
+
+type commandConfigService interface {
+	GetCommandFlagArgs() []string
+	GetCommandDir() string
+	GetCommandEnvPath() *string
+}
+
+type loggerService interface {
+	NewSlogLoggerEntry(fields ...any) *slog.Logger
+	NewSlogNamedLoggerEntry(named string, fields ...any) *slog.Logger
+	NewSlogLoggerEntryWithFields(fields ...slog.Attr) *slog.Logger
+}
+
+//nolint:interfacebloat //it's ok here, we need it we must use it as one big interface
+type errorFormatterService interface {
+	ErrorWithCode(err error, code int) error
+	ErrWithCode(err error, code int) error
+	ErrorGetCode(err error) int
+	ErrGetCode(err error) int
+	// ErrorNoWrap function for pseudo-wrap error, must be used in case of linter warnings...
+	ErrorNoWrap(err error) error
+	// ErrNoWrap same with ErrorNoWrap function, just alias for ErrorNoWrap, just short function name...
+	ErrNoWrap(err error) error
+	ErrorOnly(err error, details ...string) error
+	Error(err error, details ...string) error
+	Errorf(err error, format string, args ...interface{}) error
+	NewError(details ...string) error
+	NewErrorf(format string, args ...interface{}) error
 }